fixes issue with /etc/ssh/sdm_ca.pub permissions
Before this fix it was possible to deploy a server but not be able to connect to it, due ssh ignoring the CA file due to invalid permissions.

The file is being created via the tee -a it inherits the default permission mask, which has group/world read in most instances.

This adds a `chmod 600 /etc/ssh/sdm_ca.pub` to both templates to ensure the right permissions are enforced